Extract SyncConfig to config module
